requirements.txt
Django==4.0.6
djangorestframework==3.14.0

